pyproject.toml
[project]
name = "echi-scene-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "frozendict>=2.4.6",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "matplotlib==3.9.3",
    "mutagen==1.47.0",
    "numpy==2.0.2",
    "omegaconf==2.3.0",
    "pandas==2.2.3",
    "plotly>=5.24.1",
    "setuptools>=75.6.0",
    "soundfile==0.12.1",
    "transformers==4.47.0",
    "webrtcvad>=2.0.10",
]

[dependency-groups]
dev = [
    "ipython>=8.30.0",
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.3.241126",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "types-tqdm>=4.67.0.20241119",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
